resume-generator/pyproject.toml
[project]
name = "resume-generator"
version = "0.1.0"
description = "Resume Generator using Claude Agent SDK with custom MCP tools"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "python-docx>=1.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
